Moved base header helpers to pkg/common
