+++ pkg/epp/util/request/headers.go
@@ -19,15 +19,9 @@ package request
 import (
 	"strings"
 
-	corev3 "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
-	extProcPb "github.com/envoyproxy/go-control-plane/envoy/service/ext_proc/v3"
 	"k8s.io/apimachinery/pkg/util/sets"
 
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/metadata"
-)
-
-const (
-	RequestIdHeaderKey = "x-request-id"
 )
 
 var (
@@ -54,27 +48,4 @@ var (
 func IsSystemOwnedHeader(key string) bool {
 	k := strings.ToLower(key)
 	return InputControlHeaders.Has(k) || OutputInjectionHeaders.Has(k) || ProtocolHeaders.Has(k)
-}
-
-// GetHeaderValue safely extracts the string value from an Envoy HeaderValue field.
-func GetHeaderValue(header *corev3.HeaderValue) string {
-	if len(header.RawValue) > 0 {
-		return string(header.RawValue)
-	}
-	return header.Value
-}
-
-// ExtractHeaderValue searches for a specific header key in the processing request and returns its value.
-// The lookup is case-insensitive.
-// Returns an empty string if the header is missing or if the request structure is nil.
-func ExtractHeaderValue(req *extProcPb.ProcessingRequest_RequestHeaders, headerKey string) string {
-	headerKeyInLower := strings.ToLower(headerKey)
-	if req != nil && req.RequestHeaders != nil && req.RequestHeaders.Headers != nil {
-		for _, headerKv := range req.RequestHeaders.Headers.Headers {
-			if strings.ToLower(headerKv.Key) == headerKeyInLower {
-				return GetHeaderValue(headerKv)
-			}
-		}
-	}
-	return ""
 }
